kajonggtest: wait between terminating kajonggserver and removing socket
SVN_SILENT

svn path=/trunk/KDE/kdegames/kajongg/; revision=1280950

+++ src/kajonggtest.py
@@ -141,7 +141,9 @@ def stopServers(serverProcesses):
     """stop server processes"""
     for process, socketName in serverProcesses:
         process.terminate()
-        os.remove(socketName)
+        _ = process.wait()
+        if os.path.exists(socketName):
+            os.remove(socketName)
 
 def doJobs(jobs, options, serverProcesses):
     """now execute all jobs"""
